extracted stores listview widget
