+++ pubspec.yaml
@@ -16,7 +16,6 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   google_fonts: ^6.1.0
-  curved_navigation_bar: ^1.0.3
   currency_symbols: ^0.0.1+2
 
 dev_dependencies:
